WIP: Create separate class for Aggregate Root and move there domain events from Base Entity
