+++ CarFactory.Employees.Domain/Common/AggregateRoot.cs
@@ -1,0 +1,21 @@
+using System;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace CarFactory.Employees.Domain.Common;
+
+public abstract class AggregateRoot : BaseEntity
+{
+    private readonly List<DomainEvent> _domainEvents = new();
+    [NotMapped]
+    public IReadOnlyCollection<DomainEvent> DomainEvents => _domainEvents.AsReadOnly();
+
+    protected void RaiseDomainEvent(DomainEvent domainEvent)
+    {
+        _domainEvents.Add(domainEvent);
+    }
+
+    public void ClearDomainEvents()
+    {
+        _domainEvents.Clear();
+    }
+}
